requirement.txt
flask==3.0.0
pillow==10.2.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
torch==2.2.0
torchvision==0.17.0
